\name{wast-package}
\alias{wast-package}
\alias{wast}
\docType{package}
\title{
\packageTitle{wast}
}
\description{
\packageDescription{wast}
}
\details{
\tabular{ll}{
Package: \tab wast\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2022-05-5\cr
License: \tab GPL (>= 2)\cr
}
}
\references{
    Andrews, D. W. K. and Ploberger, W. (1994). Optimal tests when a nuisance parameter is
    present only under the alternative. Econometrica, 62(6):1383-1414.

    Fan, A., Rui, S., and Lu, W. (2017). Change-plane analysis for subgroup detection and sample
    size calculation. Journal of the American Statistical Association, 112(518):769-778.

    Huang, Y., Cho, J., and Fong, Y. (2021). Threshold-based subgroup testing in logistic regression
    models in two phase sampling designs. Journal of the Royal Statistical Society: Series C. 291-311.

    Liu, X. (2022). Subgroup testing in generalized linear models. Manuscript.
}

\keyword{ package }
